requirements.txt
Flask==3.0.2
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
werkzeug==3.0.1 